Use doubles to keep track of obj error.

+++ src/lib/pkmeans/pkmeans.h
@@ -52,7 +52,7 @@ class PKMeans {
   size_t numObservedOnce;
   float maxMissingMass;
   float confidenceProb;
-  float bestError = std::numeric_limits<float>::max();
+  double bestError = std::numeric_limits<double>::max();
 
   void readDistributions(const std::string& inFileName);
   void saveClusters(const std::string& outFilename);
@@ -80,7 +80,7 @@ class PKMeans {
   void computeNewClusters();
   void computeLowerBounds();
   void computeUpperBounds();
-  void markClustersObserved(float objError);
+  void markClustersObserved(double objError);
   void reset();
   bool needsClusterUpdateApprox(size_t x, size_t c);
   bool needsClusterUpdate(size_t x, size_t c);
@@ -92,7 +92,7 @@ class PKMeans {
   T computeDcDist(size_t x, size_t c);
   T cDist(size_t c1, size_t c2);
   T& getLowerBounds(size_t x, size_t c);
-  float calcObjFn();
+  double calcObjFn();
   static void* assignDistributionsThread(void* args);
   static void* computeClusterDistsThread(void* args);
   static void* computeNewClustersThread(void* args);
